Eliminate "progress" and "progressed"
We have always been at war with "progress".

The whole "progress" mechanism always stuck out somewhat oddly from the
rest of the design here.  It was a little bit of out-of-band data of a
weird type - a float, where every other kind of data was of an arbirary
application-specified type - and it required strange bookkeeping and
didn't really specify a clear contract.

The *purpose* of the progress method was to allow for a sort of
dual-timeout consumer: a consumer of discrete, but large objects (such
as downloaded files) where the thing that was going to be calling
receive() with your FilePath (or similar) knew when progress was being
made towards receiving a large discrete value but couldn't give the
consumer any data yet.

The idea was that for your short "wire inactivity" timeout, you could
re-set it each time "progress" was called, and for your long "no object
received" timeout you could wait for "receive" to be called.

Instead, you can handle this sort of timeout distinction at flow
construction time.  Simply do `series(fount, Timeout(wire_inactivity),
object_downloader_tube, Timeout(maximum_time_to_wait_for_an_object),
object_processing_tube, drain)`.

This eliminates a bunch of nasty edge-cases and simplifies the
interfaces down to a more comprehensible core of
started/received/stopped.

git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/branches/tubes-1956-6@42653 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb